Rename e2e dir test and move e2e tests to test/e2e
