+++ .github/workflows/e2e.yaml
@@ -77,14 +77,14 @@ jobs:
           cluster_name: kcp-cluster-1
           node_image: kindest/node:v1.21.1@sha256:fae9a58f17f18f06aeac9772ca8b5ac680ebbed985e266f711d936e91d113bad
           version: ${{ env.KIND_VERSION }}
-          config: ./e2e/kind.yaml
+          config: ./test/e2e/kind.yaml
       - name: Kubernetes KinD Cluster 2
         uses: helm/kind-action@v1.2.0
         with:
           cluster_name: kcp-cluster-2
           node_image: kindest/node:v1.21.1@sha256:fae9a58f17f18f06aeac9772ca8b5ac680ebbed985e266f711d936e91d113bad
           version: ${{ env.KIND_VERSION }}
-          config: ./e2e/kind.yaml
+          config: ./test/e2e/kind.yaml
       - name: Info
         run: |
           echo "Cluster 1:"
