+++ test/support/core.go
