+++ test/support/domainVerification.go
